requirements.txt
requests
python-telegram-bot
apscheduler
beautifulsoup4
